Refactor sox compatibility test
